+++ src/utils.tsx
@@ -80,3 +80,12 @@ export function coerceToInteger(
 
   return coerced
 }
+
+export function flattenDeep<T = any>(arr: any[]): T[] {
+  // Node.js >= 11 has supported Array.prototype.flat
+  if ((Array.prototype as any).flat) return (arr as any).flat(Infinity)
+
+  return Array.isArray(arr)
+    ? arr.reduce((a, b) => a.concat(flattenDeep(b)), [] as T[])
+    : [arr]
+}